cool#6893 reduce allocation in ScGridWindow::PaintTile
the code is needlessly allocating way more rows than we need, especially
when we have page to the bottom of a large document.

Make it so we allocate exactly the number of rows we need, instead
of allocating a default large number of rows (1024).

Which reveals that we need to allocate two extra rows, not sure why.

Change-Id: I9ca38f2712480ee8c0c3254061c92e457e328416
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162277
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>