tdf#153909 docx import: wrap through shapes shouldn't FollowTextFlow
This resolves a NISZ 6.4.0 regression from
commit 27d04f6dbf38aa28fb7215590d578c4567db5770
    tdf#119038 DOCX: fix FollowTextFlow handling

MSO is all over the place in how they implement layoutInCell
(aka FollowTextFlow in LO). There are lots of compat15 differences
for this, but I don't notice compat coming into play with this bug.

For reference, see commit e993638d5ecd33783f2eebdccfa87a81e5a8a2c5
Author: Miklos Vajna on Mon Jan 24 12:53:25 2022 +0100
    DOCX import: fix <wp:anchor layoutInCell="1"> with <wp:wrapNone>

I'm sure this is the WRONG code spot for handling FollowTextFlow,
but anywhere else and we lose the "it came from a vml shape"
limiting information, so I'm not going to try to relocate this code.

Inline (as-character) shapes ignore layoutInCell,
so the type of anchor (floating or inline) shouldn't matter
and therefore doesn't need to be part of the if () clause.

The header can be special (in compat14 it means everything
is wrap through), but at least in this case it is not limited
to IsInHeaderHeader. Using a non-header as the unit test.

make CppunitTest_sw_ooxmlexport21 \
    CPPUNIT_TEST_NAME=testTdf153909_followTextFlow

The first patchset contains a list of all of the unit tests
that were impacted by this patch: nothing interesting.

Change-Id: I0c4c7924833550533ad1b0b7609840a666d4d589
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162324
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
(cherry picked from commit ad0266eb84eafa32ccc4e0ddf3c6392860bc9b13)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162286
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>