tdf#159458: make sure to update unexpanded URL when renaming library
1. If passed new source URL is expanded (i.e., trying to expand it in
SfxLibraryContainer::checkStorageURL didn't change it), previously
aUnexpandedStorageURL was unchanged. It should be cleared - then it
will be set automatically to the URL to lib info file.

2. Try to create the "variable" path (starting with "$(USER)/basic/")
because this avoids absolute paths to the libraries, which make it
harder to move the profile.

Change-Id: Ie83ea9a32cfddf37727d238547663aeb72487ee3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162777
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 907d4107896234b2e7ebfd42f40cee3bd5b1d01e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162728
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>