tdf#159483 sc HTML import: handle data-sheets-value attribute for the bool case
Copy a cell with boolean TRUE and FALSE from google docs to Calc, the
paste result is a TRUE and a FALSE string.

The problem is that boolean is meant to be a float 0 or 1 with custom
cell format, but we don't build this doc model.

Fix the problem by ParseDataSheetsValue() to write the properties of the
cell similar to what the normal HTML import would extract from our own
markup, like:

	<td height="17" align="right" sdval="1" sdnum="1033;0;BOOLEAN">TRUE</td>

This requires passing around both the value and the numbering
properties, since the cell format just decides it's a boolean, but the
cell value will decide if it's TRUE or FALSE.

(cherry picked from commit f8c95cf93ce9ab8b9b78f3af03411d0cc2e195ba)

Change-Id: Id558ced56e02bbe24330d82c3998b047dc8febdb