tdf#159483 sc HTML paste: handle data-sheets- attributes on a span
Copy multiple cells, including a formula from google sheets to Calc, the
formula is handled as a formula by the HTML paste. Do the same for a
single cell, then only the result is pasted.

The trouble is that the data-sheets-* attributes appear on <td> elements
for multiple cells, but they appear on a <span> for a single cell.

Fix the problem by extending ScHTMLLayoutParser::ProcToken() to handle
the HtmlTokenId::SPAN_ON token and share the code between the <td> and
<span> handler, so this markup works in both cases.

Note that this is the paste handler, the no changes to the normal HTML
import are made for now.

(cherry picked from commit c0da56cb3e9f9678cae7142dee03fb706a2aebd9)

Change-Id: Id749df9062d8fcb9a2f0acd928585a304efaae28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/163296
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>