sc: Handle xls documents with unknown DRM encryption better
If we can't decrypt the document, don't set the SotStorage to
an empty SvRef. It is possible the document contains a "fallback"
un-encrypted document, which we can show. This is typical for some
Excel add-ons, which encrypt the document, but also add a part
that is un-encrypted to write the instructions on how to get the
add-on.

Change-Id: I6079786b1d0dce3a819a70e057699d0a461a5be5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/163511
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 1e1072afa3726266f83d7ef8895ed406580a3450)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/163535
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>